Added a basic trivial module syntax checker for operations.
This added dependencies to "syntax-error" and "esprima" modules.

This syntax check does not check code required from the actual module, therefore only detecting syntax errors in the operations module file.

The syntax check will have to be moved to the module deployer that will ensure that a module in the system has been checked, and will take the burden off from the runtime.