Added applications logs and fixed bug in application module loading
Applications now have logs in their root directory.
There is now a core operation that retrieves an application log: /@/core/getLog

Fixed bug in the core module retrieval that dis not look in the application directory for module files (only configutations).

The server starts as a proxy (without --app argument) and as an application process (with --app argument). The server redirects the requests to the corresponding application. Currently the application port is hardcoded (10001).