Replaced back the --app application server parameter with the application ID.
The --app application server parameter is now again the appID. The app configuration is fetched before the application server starts listening.
Added more API and internal errors using M.error API in app/model.js.
Fixed issue in M.error API that could have unpredictable behavior when a argument for M.error was a function. string.replace interpreted this as a replacing function and was executing it.